+++ speech_to_text/android/src/main/kotlin/com/csdcorp/speech_to_text/SpeechToTextPlugin.kt
@@ -18,10 +18,7 @@ import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.speech.RecognitionListener
-import android.speech.RecognitionService
-import android.speech.RecognizerIntent
-import android.speech.SpeechRecognizer
+import android.speech.*
 import android.speech.SpeechRecognizer.createOnDeviceSpeechRecognizer
 import android.speech.SpeechRecognizer.createSpeechRecognizer
 import android.util.Log
@@ -41,6 +38,7 @@ import io.flutter.plugin.common.PluginRegistry.Registrar
 import org.json.JSONArray
 import org.json.JSONObject
 import java.util.*
+import java.util.concurrent.Executors
 
 
 enum class SpeechToTextErrors {
@@ -371,14 +369,37 @@ public class SpeechToTextPlugin :
             result.success(false)
             return
         }
-        var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginContext)
-        if (null == detailsIntent) {
-            detailsIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
-            detailsIntent.setPackage("com.google.android.googlequicksearchbox")
+        if (Build.VERSION.SDK_INT >= 33 ) {
+            if ( SpeechRecognizer.isOnDeviceRecognitionAvailable(pluginContext!!)) {
+                // after much experimentation this was the only working iteration of the
+                // checkRecognitionSupport that works.
+            var recognizer = createOnDeviceSpeechRecognizer(pluginContext!!)
+            var recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
+//            var recognizer = createSpeechRecognizer(pluginContext!!)
+//            var recognizerIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
+            recognizer?.checkRecognitionSupport(recognizerIntent, Executors.newSingleThreadExecutor(),
+                object : RecognitionSupportCallback {
+                    override fun onSupportResult(recognitionSupport: RecognitionSupport) {
+                        var details = LanguageDetailsChecker( result, debugLogging )
+                        details.createResponse(recognitionSupport.supportedOnDeviceLanguages )
+                        recognizer?.destroy()
+                    }
+                    override fun onError(error: Int) {
+                        debugLog("error from checkRecognitionSupport: " + error)
+                        recognizer?.destroy()
+                    }
+                })
+            }
+        } else {
+            var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginContext)
+            if (null == detailsIntent) {
+                detailsIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
+                detailsIntent.setPackage("com.google.android.googlequicksearchbox")
+            }
+            pluginContext?.sendOrderedBroadcast(
+                    detailsIntent, null, LanguageDetailsChecker(result, debugLogging),
+                    null, Activity.RESULT_OK, null, null)
         }
-        pluginContext?.sendOrderedBroadcast(
-                detailsIntent, null, LanguageDetailsChecker(result, debugLogging),
-                null, Activity.RESULT_OK, null, null)
     }
 
     private fun notifyListening(isRecording: Boolean ) {
@@ -778,7 +799,7 @@ class LanguageDetailsChecker(flutterResult: Result, logging: Boolean ) : Broadca
         }
     }
 
-    private fun createResponse(supportedLanguages: List<String>?) {
+    public fun createResponse(supportedLanguages: List<String>?) {
         val currentLocale = Locale.getDefault()
         val localeNames = ArrayList<String>()
         localeNames.add(buildIdNameForLocale(currentLocale))